feat: add island generation & loading & builder func
What is an "Island"?
Ans: Island is a jumpable platform with collision AND collection of Props on it (not implemented yet). It is used for platforming! This is the first interactive Tree part developed
- Added Island building (Implemented in Builder & IslandManager)
- Added Island generation (Implemented in BranchGenerator & IslandGenerator)
- - Algorithm (recursive generation) (simplified):
- - TreeGenerator generates Chunks using ChunkGenerator
- - ChunkGenerator generates Branches using BranchGenerator
- - BranchGenerator generates Islands using IslandGenerator
- - - Islands are generated with random allowed size
- - - When Island's size < allowed size, it gets a random offset based on delta sizes
- - Result: 1 Chunk -> 4 Branches -> 8 Islands (1-3 per Branch)
- Added Island loading (Implemented in IslandLoader & BranchLoader)
- Implemented Island in BlankLibrary

- Replaced 2 Render Layers "Sprite" & "Outline" with 3 specific "Trunk" & "Island" & "Shape"
- Added resources for testing:
- - 7 IslandData (2 Tiny, 3 Small, 2 Medium)

Note: Shortly after that update Character should arise!

+++ Assets/_Project/Scripts/Data/BranchData.cs
@@ -43,7 +43,7 @@ public class BranchData : ScriptableObject, IData
     // Unique identifier
     public string id;
     public AvailableSide avaliableSide = AvailableSide.Both;
-    public int length;
+    public float length;
 
     [Header("Islands Configuration")]
     public IslandSlot[] islandSlots;
